+++ src/code-generation/renderers/causa-validator-renderer.ts
@@ -1,4 +1,4 @@
-import { removeNullFromType } from 'quicktype-core/dist/TypeUtils.js';
+import { removeNullFromType } from 'quicktype-core/dist/Type/index.js';
 import type { TypeScriptDecorator } from '../decorator.js';
 import {
   type ClassPropertyContext,